requirements.txt
aiohttp
discord.py
parsedatetime
python-dateutil
PyYAML
typing_extensions